HPCC-31375 Exceeded dafilesrv CSV/XML streaming replyLimit fix
If a streaming dafilesrv client read a file that exceeded the
replyLimit on a CSV or XML file, it would try to serialize the
cursor, which has no implemention. This resulted in an exception
being reported back to the client.
The cursor is only used if the client needs to restore the
connection (on a different instance of a dafilesrv),
i.e. it is not used for the vast majority of use cases.
Instead of throwing an exception during cursor serialization,
serialize a dummy 'UNSUPPORTED' placeholder.
If there is any attempt to use this placeholder, i.e. during
continuation, the deseriailze call in dafilesrv will throw an
exception.

+++ system/jlib/jptree.cpp
@@ -6608,7 +6608,7 @@ void printJSON(const IPropertyTree *tree, unsigned indent, byte flags)
     printf("%s", json.str());
 }
 
-void dbglogJSON(const IPropertyTree *tree, unsigned indent, byte flags)
+void dbglogJSON(const IPropertyTree *tree, unsigned indent, unsigned flags)
 {
     StringBuffer json;
     toJSON(tree, json, indent, flags);
